backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0

# Testing
pytest>=7.4.0
httpx>=0.26.0
pytest-asyncio>=0.23.0
